requirements.txt
asyncio
pip
psycopg2-binary
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
